chore: StepPages -> StepFunnel rename
