+++ src/common/components/StepIndicator/index.tsx
