+++ src/common/components/StepIndicator/style.css.ts
