+++ src/features/post/components/StepFunnel/step1.tsx
