+++ src/features/post/components/StepFunnel/step2.tsx
