+++ src/features/post/components/StepFunnel/step6.tsx
